+++ toys-core/test/test_dsl.rb
@@ -133,6 +133,7 @@ describe Toys::DSL::Tool do
       end
       tool, _remaining = loader.lookup(["foo"])
       assert_equal(false, tool.runnable?)
+      assert_equal(false, tool.interruptable?)
     end
 
     it "makes a tool runnable when the run method is defined" do
@@ -164,6 +165,27 @@ describe Toys::DSL::Tool do
       end
       tool, _remaining = loader.lookup(["foo"])
       assert_equal(true, tool.runnable?)
+    end
+
+    it "makes a tool interruptable when the interrupt method is defined" do
+      loader.add_block do
+        tool "foo" do
+          def interrupt; end
+        end
+      end
+      tool, _remaining = loader.lookup(["foo"])
+      assert_equal(true, tool.interruptable?)
+    end
+
+    it "makes a tool interruptable when the to_interrupt directive is given" do
+      loader.add_block do
+        tool "foo" do
+          to_interrupt do
+          end
+        end
+      end
+      tool, _remaining = loader.lookup(["foo"])
+      assert_equal(true, tool.interruptable?)
     end
 
     it "makes callable methods" do